requirements.txt
asttokens==2.4.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.0.7
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.2
fonttools==4.50.0
ipykernel==6.29.4
ipython==8.12.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.24.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
parso==0.8.3
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pyzmq==25.1.2
six==1.16.0
stack-data==0.6.3
tornado==6.4
traitlets==5.14.2
typing_extensions==4.10.0
wcwidth==0.2.13
Werkzeug==3.0.1
pywin32-ctypes==0.2.0
gunicorn